+++ HISTORY.rst
@@ -3,6 +3,24 @@
 Release History
 ---------------
 
+2.9.0 (2020-10-12)
++++++++++++++++++++
+
+**Improvements**
+
+- Fix types for list_race_details (synapticarbors)
+- #340 cache removal added on old markets (8 hours closed)
+- Streaming snap added to resources
+
+**Bug Fixes**
+
+- PR added to actions
+
+**Dependencies**
+
+- py3.9 added to tests
+- orjson updated to 3.4.0
+
 2.8.0 (2020-09-14)
 +++++++++++++++++++